Revised README.md
The team updated various sections of the README to ensure that the document was more concise, had the necessary embedded links, and added links for additional for access to the mailing list, zulip chat, and the community meetings. 

The team updated the README to align it with GitHub Best Practices and mimic successful large scale repositories that are widely know for their impeccable documentation. 

Signed-off-by: nserway <146010453+nserway@users.noreply.github.com>